Refactor fight packages
Add fight.game package to put the elements of the game stage.
Add fight.popup package to put the elements of the popup menu stage.

Issue #63
