+++ core/src/fr/mmyumu/troncgame/components/FightComponent.java
@@ -2,12 +2,9 @@ package fr.mmyumu.troncgame.components;
 
 import dagger.Component;
 import fr.mmyumu.troncgame.ActivityScope;
-import fr.mmyumu.troncgame.fight.FightGame;
-import fr.mmyumu.troncgame.fight.FightGameInputProcessor;
+import fr.mmyumu.troncgame.fight.game.FightGameInputProcessor;
 import fr.mmyumu.troncgame.fight.FightLoadingScreen;
-import fr.mmyumu.troncgame.fight.FightPopUpMenu;
 import fr.mmyumu.troncgame.fight.FightScreen;
-import fr.mmyumu.troncgame.fight.ui.FightUI;
 import fr.mmyumu.troncgame.modules.FightModule;
 
 /**
