+++ core/src/fr/mmyumu/troncgame/fight/popup/FightPopUpMenu.java
@@ -1,4 +1,4 @@
-package fr.mmyumu.troncgame.fight;
+package fr.mmyumu.troncgame.fight.popup;
 
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.utils.viewport.ScalingViewport;
