+++ core/src/fr/mmyumu/troncgame/fight/popup/FightPopUpMenuIcon.java
@@ -1,4 +1,4 @@
-package fr.mmyumu.troncgame.fight;
+package fr.mmyumu.troncgame.fight.popup;
 
 import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.Texture;
