+++ core/src/fr/mmyumu/troncgame/fight/popup/FightPopUpMenuLogic.java
@@ -1,4 +1,4 @@
-package fr.mmyumu.troncgame.fight;
+package fr.mmyumu.troncgame.fight.popup;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -6,26 +6,28 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import fr.mmyumu.troncgame.fight.*;
+
 /**
  * Manage the logic of the pop up menu
  * Created by mmyumu on 22/01/2016.
  */
 public class FightPopUpMenuLogic {
-    private List<FightPopUpMenuIcon> popMenuIcons;
+    private List<fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon> popMenuIcons;
     private FightPopUpMenuNotReady popUpMenuNotReady;
 
     @Inject
-    public FightPopUpMenuLogic(FightPopUpMenuNotReady fightPopUpMenuNotReady, @Named("spells") FightPopUpMenuIcon fightPopUpMenuSpellsIcon, @Named("weapons") FightPopUpMenuIcon fightPopUpMenuWeaponsIcon) {
+    public FightPopUpMenuLogic(FightPopUpMenuNotReady fightPopUpMenuNotReady, @Named("spells") fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon fightPopUpMenuSpellsIcon, @Named("weapons") fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon fightPopUpMenuWeaponsIcon) {
         this.popUpMenuNotReady = fightPopUpMenuNotReady;
 
-        popMenuIcons = new ArrayList<FightPopUpMenuIcon>();
+        popMenuIcons = new ArrayList<fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon>();
         popMenuIcons.add(fightPopUpMenuSpellsIcon);
         popMenuIcons.add(fightPopUpMenuWeaponsIcon);
     }
 
     public List<FightPopUpMenuElement> getActors() {
         List<FightPopUpMenuElement> actors = new ArrayList<FightPopUpMenuElement>();
-        for (FightPopUpMenuIcon popUpMenuIcon : popMenuIcons) {
+        for (fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon popUpMenuIcon : popMenuIcons) {
             actors.add(popUpMenuIcon);
         }
 
@@ -38,14 +40,14 @@ public class FightPopUpMenuLogic {
     }
 
     public void selectReadyCharacter(FightCharacter selectedCharacter) {
-        for (FightPopUpMenuIcon fightPopUpMenuIcon : popMenuIcons) {
+        for (fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon fightPopUpMenuIcon : popMenuIcons) {
             fightPopUpMenuIcon.select();
             fightPopUpMenuIcon.display(selectedCharacter.getCenter());
         }
     }
 
-    public void selectIcon(FightPopUpMenuIcon selectedIcon) {
-        for (FightPopUpMenuIcon popUpMenuIcon : popMenuIcons) {
+    public void selectIcon(fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon selectedIcon) {
+        for (fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon popUpMenuIcon : popMenuIcons) {
             if (popUpMenuIcon.equals(selectedIcon)) {
                 popUpMenuIcon.select();
             } else {
@@ -55,7 +57,7 @@ public class FightPopUpMenuLogic {
     }
 
     public void unselectCharacter() {
-        for (FightPopUpMenuIcon fightPopUpMenuIcon : popMenuIcons) {
+        for (fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon fightPopUpMenuIcon : popMenuIcons) {
             fightPopUpMenuIcon.hide();
         }
     }
