+++ core/src/fr/mmyumu/troncgame/modules/FightModule.java
@@ -17,13 +17,13 @@ import fr.mmyumu.troncgame.Constants;
 import fr.mmyumu.troncgame.TroncGame;
 import fr.mmyumu.troncgame.fight.FightBackground;
 import fr.mmyumu.troncgame.fight.FightConstants;
-import fr.mmyumu.troncgame.fight.FightGame;
+import fr.mmyumu.troncgame.fight.game.FightGame;
 import fr.mmyumu.troncgame.fight.FightLoadingScreen;
 import fr.mmyumu.troncgame.fight.FightLogic;
-import fr.mmyumu.troncgame.fight.FightPopUpMenu;
-import fr.mmyumu.troncgame.fight.FightPopUpMenuIcon;
-import fr.mmyumu.troncgame.fight.FightPopUpMenuLogic;
-import fr.mmyumu.troncgame.fight.FightPopUpMenuNotReady;
+import fr.mmyumu.troncgame.fight.popup.FightPopUpMenu;
+import fr.mmyumu.troncgame.fight.popup.FightPopUpMenuIcon;
+import fr.mmyumu.troncgame.fight.popup.FightPopUpMenuLogic;
+import fr.mmyumu.troncgame.fight.popup.FightPopUpMenuNotReady;
 import fr.mmyumu.troncgame.fight.FightScreen;
 import fr.mmyumu.troncgame.fight.enemy.EnemyFightTeamGenerator;
 import fr.mmyumu.troncgame.fight.ui.FightMainInfos;
